Recover two lost exact functions

+++ src/town.cpp
@@ -1594,7 +1594,9 @@ unsigned char town::canBuild(short buildingId) const
 }
 
 VA(0x005c0e60, 0xC0)  // dc 0x16865c
-unsigned char town::canEverBuild(short buildingId) const
+// Complete reads the full dword parameter and its exact symbol encodes int;
+// Dreamcast's older interface records short building_id.
+unsigned char town::canEverBuild(int buildingId) const
 {
     if (g_bitNumber[buildingId] & m_available) {
         if (buildingId == DOCK_ID)
